requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
python-dotenv==1.0.1
dj-database-url==2.1.0
environs==11.0.0
